Unfinished

+++ apps/api/src/modules/subjects/domain/port/input/subjects.service.ts
@@ -8,12 +8,16 @@ import {
 import { ISubjectsService } from './ISubjectsService';
 import { PaginatedResponse } from 'src/shared/types/PaginatedResponse';
 import { SubjectsDto } from '../../dto/subjects.dto';
+import { IStudentClassEnrollRepository } from 'src/modules/student-class-enrolls/domain/port/output/IStudentClassEnrollRepository';
+import { IClassesRepository } from 'src/modules/classes/domain/port/output/IClassesRepository';
 
 @Injectable()
 export class SubjectsService implements ISubjectsService {
   constructor(
     @Inject(SUBJECTS_REPOSITORY)
     private subjectsRepository: ISubjectsRepository,
+    private studentClassEnrollRepository: IStudentClassEnrollRepository,
+    private classesRepository: IClassesRepository,
   ) {}
 
   async count(whereOptions: any): Promise<number> {
@@ -34,7 +38,27 @@ export class SubjectsService implements ISubjectsService {
 
   async delete(subjectId: string) {
     try {
-      return await this.subjectsRepository.delete(subjectId);
+      const subject = await this.subjectsRepository.findById(subjectId);
+
+      const current = new Date();
+      const thirtyMinutesAgo = new Date(current.getTime() - 30 * 60 * 1000);
+      if (subject.createdAt && subject.createdAt <= thirtyMinutesAgo) {
+        throw new Error(
+          `Cannot delete subject with ID ${subjectId} because it was created more than 30 minutes ago`,
+        );
+      } else {
+        const classes = await this.classesRepository.findBySubjectCode(
+          subject.code,
+        );
+
+        if (classes.length > 0) {
+          return this.subjectsRepository.update(subjectId, {
+            status: 'DEACTIVATED',
+          });
+        }
+
+        return await this.subjectsRepository.delete(subjectId);
+      }
     } catch (error) {
       throw new Error(
         `Error deleting subject with ID ${subjectId}: ${error.message}`,
@@ -95,6 +119,16 @@ export class SubjectsService implements ISubjectsService {
 
   async update(subjectId: string, data: UpdateSubjectDTO) {
     try {
+      const subject = await this.subjectsRepository.findById(subjectId);
+      const studentClassEnrolls =
+        await this.studentClassEnrollRepository.findBySubjectCode(subject.code);
+
+      if (subject.credit !== data.credit && studentClassEnrolls.length > 0) {
+        throw new Error(
+          `Cannot update credit of subject with ID ${subjectId} because it has student class enrollments`,
+        );
+      }
+
       return await this.subjectsRepository.update(subjectId, data);
     } catch (error) {
       throw new Error(
